requirements.txt
certifi==2020.11.8
chardet==3.0.4
colorama==0.4.4
cycler==0.10.0
idna==2.10
kiwisolver==1.3.1
lxml==4.6.1
matplotlib==3.2.2
multitasking==0.0.9
numpy==1.19.4
pandas==1.1.4
pandas-datareader==0.9.0
Pillow==8.0.1
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.4
pywin32==300
requests==2.25.0
six==1.15.0
tqdm==4.54.0
tqdm-multiprocess==0.0.11
urllib3==1.26.2
xing-plus==1.0.2
yfinance==0.1.55
